funcionalidades implementados
