+++ src/app/page.tsx
@@ -1,12 +1,11 @@
 
-import { TrabajosRecientes } from './Home/TrabajosRecientes';
+import { TrabajosRecientes } from '../componets/TrabajosRecientes';
 
 export default function Home() {
   return (
     <div>
-      
       <TrabajosRecientes />
-
+      
     </div>
   );
 }
