+++ src/componets/header.jsx
@@ -1,4 +1,5 @@
-const Header = () => {
+
+export const Header = () => {
   return (
     <header className="bg-white shadow-md">
       <div className="container mx-auto px-4 py-4 flex justify-between items-center">
